requirements.txt
langchain
langchain-huggingface
langchain-chroma
langchain-community
langchain-text-splitters
pdfminer.six
python-dotenv
chromadb